split up oracle into multiple files
