+++ packages/sdk/tasks/oracle/index.ts
@@ -1,0 +1,4 @@
+import "./add-tokens";
+import "./get-price";
+import "./set-price";
+import "./update-twap";
